add ResourceDataset handler breakup resource module
