+++ reV/handlers/__init__.py
@@ -4,5 +4,6 @@ Sub-package of data handlers
 """
 from .collection import Collector
 from .multi_year import MultiYear
-from .resource import (NSRDB, MultiFileNSRDB, MultiFileWTK, Resource,
-                       SolarResource, WindResource)
+from .resource import Resource
+from .rev_resource import (NSRDB, MultiFileNSRDB, MultiFileWTK,
+                           SolarResource, WindResource)
