+++ reV/handlers/parse_keys.py
@@ -4,6 +4,26 @@ Handler utility to parse slicing keys.
 """
 
 
+def parse_slice(ds_slice):
+    """
+    Parse dataset slice
+
+    Parameters
+    ----------
+    ds_slice : tuple | int | slice | list
+        Slice to extract from dataset
+
+    Returns
+    -------
+    ds_slice : tuple
+        slice for axis (0, 1)
+    """
+    if not isinstance(ds_slice, tuple):
+        ds_slice = (ds_slice,)
+
+    return ds_slice
+
+
 def parse_keys(keys):
     """
     Parse keys for complex __getitem__ and __setitem__
