+++ tests/test_handlers_res_hh.py
@@ -5,7 +5,7 @@ pytests for resource handlers with a single hub height
 import numpy as np
 import os
 import pytest
-from reV.handlers.resource import WindResource
+from reV.handlers.rev_resource import WindResource
 from reV import TESTDATADIR
 from reV.config.project_points import ProjectPoints
